Fix test

+++ app/src/main/java/com/daniil/shevtsov/idle/main/data/resource/ResourceDto.kt
@@ -1,5 +1,5 @@
 package com.daniil.shevtsov.idle.main.data.resource
 
 data class ResourceDto(
-    val value: Long,
+    val value: Double,
 )
