+++ app/src/main/java/com/daniil/shevtsov/idle/main/domain/resource/Resource.kt
@@ -1,5 +1,5 @@
 package com.daniil.shevtsov.idle.main.domain.resource
 
 data class Resource(
-    val value: Long,
+    val value: Double,
 )
